only shows staff information for respective user
