--- pro-dental-care-ui/src/app/admin-staff info/staff-information.component.css
@@ -1,105 +1,0 @@
-/* Wrapper for the overall layout for Staff Information */
-.content-wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  .section {
-    border: 1px solid #e0e0e0;
-    border-radius: 8px;
-    margin-bottom: 10px;
-    cursor: pointer;
-    width: 80%;
-    overflow: hidden;
-    max-height: 500px;
-    margin-left: auto;
-  }
-  
-  .section.expanded {
-    max-height: 800px;
-  }
-  
-  .section-header {
-    padding: 15px;
-    background-color: #f5f5f5;
-    font-weight: bold;
-    border-bottom: 1px solid #ddd;
-    text-align: left;
-    text-transform: uppercase;
-  }
-  
-  .section-content {
-    padding: 15px;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-  }
-  
-  .label {
-    color: #888888;
-    font-weight: normal;
-  }
-  
-  .value {
-    color: black;
-    font-weight: bold;
-  }
-  
-  .row {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-between;
-    width: 100%;
-    margin-top: 10px;
-    gap: 20px;
-  }
-  
-  .edit-icon {
-    width: 34px;
-    height: 34px;
-    cursor: pointer;
-    margin-left: 5px;
-  }
-  
-  .icon-container {
-    display: flex;
-    justify-content: flex-end;
-    width: 100%;
-    margin-top: 10px;
-  }
-  
-  .save-icon, .cancel-icon, .edit-icon {
-    margin-left: 10px;
-    cursor: pointer;
-  }
-  
-  .save-icon {
-    width: 34px;
-    height: 34px;
-  }
-  
-  .cancel-icon {
-    width: 34px;
-    height: 34px;
-  }
-
-  .info-row {
-    margin-bottom: 15px; /* Add space between rows for better readability */
-  }
-  
-  .input-with-label {
-    display: flex;
-    align-items: center;
-  }
-  
-  .input-with-label input {
-    flex: 1;
-  }
-  
-  .rate-label {
-    margin-left: 8px;
-    font-weight: bold;
-    color: #444;
-  }
-
