--- pro-dental-care-ui/src/app/admin-staff info/staff-information.component.html
@@ -1,52 +1,0 @@
-<div class="content-wrapper">
-  <!-- Dynamic Accordions for Staff Information -->
-  <div *ngFor="let staff of staffInfo; let i = index" class="section" (click)="toggleSection(i)" [class.expanded]="expandedSection === i">
-    <div class="section-header">{{ staff.name }}</div>
-    <div class="section-content" *ngIf="expandedSection === i" (click)="preventToggle($event)">
-      <div class="info-row">
-        <span class="label">Title</span><br>
-        <span *ngIf="!editMode" class="value">{{ staff.title }}</span>
-        <input *ngIf="editMode" type="text" [(ngModel)]="staff.title" />
-      </div>
-      <div class="info-row">
-        <span class="label">Supervisor</span><br>
-        <span *ngIf="!editMode" class="value">{{ staff.supervisor }}</span>
-        <input *ngIf="editMode" type="text" [(ngModel)]="staff.supervisor" />
-      </div>
-      <div class="info-row">
-        <span class="label">Staff ID</span><br>
-        <span *ngIf="!editMode" class="value">{{ staff.staffId }}</span>
-        <input *ngIf="editMode" type="text" [(ngModel)]="staff.staffId" />
-      </div>
-      <div class="row">
-        <div class="info-row">
-          <span class="label">Employment Type</span><br>
-          <span *ngIf="!editMode" class="value">{{ staff.employmentType }}</span>
-          <select *ngIf="editMode" [(ngModel)]="staff.employmentType">
-            <option value="Full-Time">Full-Time</option>
-            <option value="Part-Time">Part-Time</option>
-            <option value="Contractor">Contractor</option>
-          </select>
-        </div>
-        <div class="info-row">
-          <span class="label">Pay Rate</span><br>
-          <span *ngIf="!editMode" class="value">{{ staff.payRate }} $/hour</span>
-          <div *ngIf="editMode" class="input-with-label">
-            <input type="number" [(ngModel)]="staff.payRate" step="0.01" min="0" />
-            <span class="rate-label">$/hour</span>
-          </div>
-        </div>
-        <div class="info-row">
-          <span class="label">Years of Service</span><br>
-          <span *ngIf="!editMode" class="value">{{ staff.yearsOfService }}</span>
-          <input *ngIf="editMode" type="number" [(ngModel)]="staff.yearsOfService" />
-        </div>
-      </div>
-      <div class="icon-container">
-        <img *ngIf="!editMode" class="edit-icon" ngSrc="/assets/edit_icon.png" width="34" height="34" (click)="enableEdit($event)" />
-        <img *ngIf="editMode" class="save-icon" ngSrc="/assets/save_icon.png" width="34" height="34" (click)="saveButton($event)" />
-        <img *ngIf="editMode" class="cancel-icon" ngSrc="/assets/cancel_icon.png" width="34" height="34" (click)="cancelButton($event)" />
-      </div>
-    </div>
-  </div>
-</div>
