+++ pro-dental-care-ui/src/app/staff-staff info/staff-information.component.css
@@ -1,0 +1,30 @@
+/* Wrapper for the overall layout */
+.content-wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px;
+}
+
+/* Centered Name Header */
+.staff-name {
+  font-size: 24px;
+  font-weight: bold;
+  text-align: center;
+  margin-bottom: 20px;
+}
+
+/* Staff Information Section */
+.section {
+  border: 1px solid #e0e0e0;
+  border-radius: 8px;
+  width: 100%;
+  max-width: 600px;
+  padding: 20px;
+  background-color: #f9f9f9;
+}
+
+.section-content p {
+  margin: 10px 0;
+  font-size: 16px;
+}
