+++ pro-dental-care-ui/src/app/staff-staff info/staff-information.component.html
@@ -1,0 +1,16 @@
+<div class="content-wrapper">
+  <!-- Staff Name Header -->
+  <h2 class="staff-name">{{ staffInfo.name }}</h2>
+
+  <!-- Staff Information Section -->
+  <div class="section">
+    <div class="section-content">
+      <p><strong>Title:</strong> {{ staffInfo.title }}</p>
+      <p><strong>Supervisor:</strong> {{ staffInfo.supervisor }}</p>
+      <p><strong>Staff ID:</strong> {{ staffInfo.staffId }}</p>
+      <p><strong>Employment Type:</strong> {{ staffInfo.employmentType }}</p>
+      <p><strong>Pay Rate:</strong> ${{ staffInfo.payRate }}/hour</p>
+      <p><strong>Years of Service:</strong> {{ staffInfo.yearsOfService }}</p>
+    </div>
+  </div>
+</div>
